Correction globale

+++ scraper/main.py
@@ -1,30 +1,32 @@
 import pandas as pd
 import time
-from indeed import IndeedScraper
+from indeed.IndeedScraper import IndeedScraper
 from wttj.WTTJScraper import WTTJScraper
 from threading import Thread
+import os
 
 from Data_Cleaner import DataCleaner
 
-ISc = IndeedScraper(
+"""ISc = IndeedScraper(
     webdriver_path='./webdriver/chromedriver.exe',
-    nb_pages=5,
+    nb_pages=2,
     output_name="./data/INDEED.csv",
     min_delai=3,
-    update_every=0
-)
+    update_every=1
+)"""
 
 wts = WTTJScraper(
     webdriver_path='./webdriver/chromedriver.exe',
-    nb_pages=5,
+    nb_pages=2,
     output_name="./data/WTTJ.csv",
-    update_every=0,
+    url_csv='./data/temp/temp_wttj_links.csv',
+    update_every=1,
 )
 
 thread_list = []
-thrd1 = Thread(target=ISc.launch_scraping)
+#thrd1 = Thread(target=ISc.launch_scraping)
 thrd2 = Thread(target=wts.launch_scraping)
-thread_list.append(thrd1)
+#thread_list.append(thrd1)
 thread_list.append(thrd2)
 
 # Demarre le thread
@@ -42,7 +44,14 @@ time.sleep(3)
 df1 = pd.read_csv("./data/INDEED.csv")
 df2 = pd.read_csv("./data/WTTJ.csv")
 df3 = pd.concat([df1, df2], axis=0, ignore_index=True)
-df3.to_csv('./data/Ind_Wttj.csv', index=False)
+
+if os.path.isfile('./data/Ind_Wttj.csv')==True :
+    df4 = pd.read_csv('./data/Ind_Wttj.csv')
+    df_final = pd.concat([df4,df3],axis=0,ignore_index=True)
+    df_final.to_csv('./data/Ind_Wttj.csv',index=False)
+else :
+    df3.to_csv('./data/Ind_Wttj.csv', index=False)
+
 
 # Clean csv and export the result
 dc = DataCleaner('./data/Ind_Wttj.csv')
